+++ apparat_launcher/plugin_nautilus.py
@@ -1,5 +1,5 @@
 #!/usr/bin/python
-"""apparat_launcher - plugin: nautilus"""
+"""plugin: nautilus (optional)"""
 
 # general
 import os
@@ -19,9 +19,9 @@ TRIGGER = ('!goto', '!recent', '!trash', '!network', '!net',)
 # FUNCTIONS
 # -----------------------------------------------------------------------------------------------
 
-def prepare_general(current_search_string, main_window):
-    """Prepare General"""
-    tools.debug_output(__name__, 'prepare_general', 'starting', 1)
+def parse(current_search_string, main_window):
+    """Validates the input and calls the matching sub-function"""
+    tools.debug_output(__name__, 'parse', 'starting', 1)
 
     # Reset status notification back to OK
     main_window.status_notification_reset()
@@ -29,30 +29,30 @@ def prepare_general(current_search_string, main_window):
     icon_size = ini.read_single_ini_value('General', 'icon_size') # get preference value
 
     if current_search_string.startswith('!goto'):
-        tools.debug_output(__name__, 'prepare_general', 'Case: Goto', 1)
+        tools.debug_output(__name__, 'parse', 'Case: Goto', 1)
         prepare_plugin_nautilus_goto(main_window, icon_size)
 
     elif current_search_string == ('!recent'):
-        tools.debug_output(__name__, 'prepare_general', 'Case: Recent', 1)
+        tools.debug_output(__name__, 'parse', 'Case: Recent', 1)
         prepare_plugin_nautilus_show_recent(main_window, icon_size)
 
     elif current_search_string == ('!trash'):
-        tools.debug_output(__name__, 'prepare_general', 'Case: Trash', 1)
+        tools.debug_output(__name__, 'parse', 'Case: Trash', 1)
         prepare_plugin_nautilus_open_trash(main_window, icon_size)
 
     elif current_search_string == ('!network') or current_search_string == ('!net'):
-        tools.debug_output(__name__, 'prepare_general', 'Case: Network', 1)
+        tools.debug_output(__name__, 'parse', 'Case: Network', 1)
         prepare_plugin_nautilus_show_network_devices(main_window, icon_size)
 
     else:
-        tools.debug_output(__name__, 'prepare_general', 'Error: Unexpected nautilus plugin command', 3)
+        tools.debug_output(__name__, 'parse', 'Error: Unexpected nautilus plugin command', 3)
         main_window.status_notification_display_error('Unexpected nautilus plugin command')
 
-    tools.debug_output(__name__, 'prepare_general', 'finished', 1)
+    tools.debug_output(__name__, 'parse', 'finished', 1)
 
 
 def prepare_plugin_nautilus_goto(main_window, icon_size):
-    """Plugin Nautilus - GoTo"""
+    """Plugin Nautilus - GoTo - Opens a path in nautilus"""
     tools.debug_output(__name__, 'prepare_plugin_nautilus_goto', 'starting', 1)
 
     ## update plugin info
@@ -78,7 +78,7 @@ def prepare_plugin_nautilus_goto(main_window, icon_size):
 
 
 def prepare_plugin_nautilus_show_network_devices(main_window, icon_size):
-    """Plugin Nautilus - Network"""
+    """Plugin Nautilus - Network - Opens network view in nautilus"""
     tools.debug_output(__name__, 'prepare_plugin_nautilus_show_network_devices', 'starting', 1)
 
     ## update plugin info
@@ -98,7 +98,7 @@ def prepare_plugin_nautilus_show_network_devices(main_window, icon_size):
 
 
 def prepare_plugin_nautilus_show_recent(main_window, icon_size):
-    """Plugin Nautilus - Recent"""
+    """Plugin Nautilus - Recent - Opens recent files in nautilus"""
     tools.debug_output(__name__, 'prepare_plugin_nautilus_show_recent', 'starting', 1)
 
     ## update plugin info
@@ -118,7 +118,7 @@ def prepare_plugin_nautilus_show_recent(main_window, icon_size):
 
 
 def prepare_plugin_nautilus_open_trash(main_window, icon_size):
-    """Plugin Nautilus - Trash"""
+    """Plugin Nautilus - Trash - Opens trash in nautilus"""
     tools.debug_output(__name__, 'prepare_plugin_nautilus_open_trash', 'starting', 1)
 
     ## update plugin info