+++ apparat_launcher/plugin_shell.py
@@ -1,5 +1,5 @@
 #!/usr/bin/python
-"""apparat_launcher - plugin: shell"""
+"""plugin: shell (optional)"""
 
 ## general
 import wx
@@ -19,14 +19,14 @@ TRIGGER = ('!sh',)
 # -----------------------------------------------------------------------------------------------
 # FUNCTIONS
 # -----------------------------------------------------------------------------------------------
-def prepare_general(current_search_string, main_window):
-    """Prepare General"""
-    tools.debug_output(__name__, 'prepare_general', 'starting', 1)
+def parse(current_search_string, main_window):
+    """parse"""
+    tools.debug_output(__name__, 'parse', 'starting', 1)
     main_window.status_notification_reset() # Reset status notification back to OK
     icon_size = ini.read_single_ini_value('General', 'icon_size') # get preference value
 
     if current_search_string.startswith("!sh"):
-        tools.debug_output(__name__, 'prepare_general', 'Case: Shell', 1)
+        tools.debug_output(__name__, 'parse', 'Case: Shell', 1)
         prepare_plugin_shell(main_window, icon_size)
 
         if (len(current_search_string) > 3):
@@ -34,15 +34,13 @@ def prepare_general(current_search_string, main_window):
                 prepare_plugin_shell(main_window, icon_size)
             else:
                 main_window.plugin__update_general_ui_information('')
-                tools.debug_output(__name__, 'prepare_general', 'Aborting shell', 2)
+                tools.debug_output(__name__, 'parse', 'Aborting shell', 2)
         return
 
     else:
-        tools.debug_output(__name__, 'prepare_general', 'Error: Unexpected shell plugin command', 3)
+        tools.debug_output(__name__, 'parse', 'Error: Unexpected shell plugin command', 3)
         main_window.status_notification_display_error('Unexpected shell plugin command')
         return
-
-    tools.debug_output(__name__, 'prepare_general', 'finished', 1)
 
 
 def prepare_plugin_shell(main_window, icon_size):